pyproject.toml
[project]
name = "laintsxwin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-rich-presence>=1.1.0",
    "pywebview>=6.0",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.15.0",
]
